fix(tinkerbell): stop template action details from bleeding across actions
parseTaskActions read an action's image/timeout with a fixed-size line
window (8 lines for direct actions, 10 past the end marker for
conditional ones). When an action's own definition was shorter than
that window, the search continued into the next action's lines and
could attribute its image or timeout to the wrong action instead of
falling back to "no value".

Bound the search to the next action or control marker instead of a
fixed line count, so each action's details come only from its own
lines regardless of how short or long it is.

Move the template parsing helpers out of Detail.tsx into helpers.ts
(matching the workflows/helpers.ts pattern already used in this
plugin) so they're unit-testable, and add tests covering the bleed
case, a long multi-line action, and the existing task/conditional
parsing behavior.
